+++ islands/DemoFooter.tsx
@@ -1,0 +1,6 @@
+// import { useState } from "preact/hooks";
+import { Footer } from "lunchbox/components.ts";
+
+export default function () {
+  return <Footer></Footer>;
+}