+++ islands/DemoHeader.tsx
@@ -1,0 +1,6 @@
+// import { useState } from "preact/hooks";
+import { Header } from "lunchbox/components.ts";
+
+export default function () {
+  return <Header></Header>;
+}