+++ islands/DemoMain.tsx
@@ -1,0 +1,6 @@
+// import { useState } from "preact/hooks";
+import { Main } from "lunchbox/components.ts";
+
+export default function () {
+  return <Main></Main>;
+}